+++ vrc-get-vpm/src/environment.rs
@@ -4,8 +4,12 @@ mod settings;
 mod uesr_package_collection;
 mod vrc_get_settings;
 
+#[cfg(feature = "vrc-get-litedb")]
+mod litedb;
 #[cfg(feature = "experimental-project-management")]
 mod project_management;
+#[cfg(feature = "experimental-unity-management")]
+mod unity_management;
 
 use crate::io;
 use crate::io::SeekFrom;
@@ -55,8 +59,8 @@ pub struct Environment<T: HttpClient, IO: EnvironmentIo> {
     vrc_get_settings: VrcGetSettings,
     // we do not connect to litedb unless we need information from litedb.
     // TODO?: use inner mutability?
-    #[cfg(feature = "experimental-project-management")]
-    litedb_connection: Option<vrc_get_litedb::DatabaseConnection>,
+    #[cfg(feature = "vrc-get-litedb")]
+    litedb_connection: litedb::LiteDbConnectionHolder,
     /// Cache
     repo_cache: RepoHolder,
     user_packages: UserPackageCollection,
@@ -69,7 +73,7 @@ impl<T: HttpClient, IO: EnvironmentIo> Environment<T, IO> {
             settings: Settings::load(&io).await?,
             vrc_get_settings: VrcGetSettings::load(&io).await?,
             #[cfg(feature = "vrc-get-litedb")]
-            litedb_connection: None,
+            litedb_connection: litedb::LiteDbConnectionHolder::new(),
             repo_cache: RepoHolder::new(),
             user_packages: UserPackageCollection::new(),
             io,